+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-01-27
+
+### Fixed
+- The Web site has changed some component xpath.
+
+### Changed
+- Upgrade Selenium version to 4.17.2
+- Upgrade Geckodriver version to 0.34.0
+
+## [0.2.1] - 2023-02-05
+
+### Fixed
+- The Web site has changed some component xpath. 
+
 ## [0.2.0] - 2022-10-17
 
 ### Added